Integration testing changes
Refactoring. Some functions moved to separate utils module. Some documentation and comments added.
